# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Patchman-engine API",
        "description": "API of the Patch application on [console.redhat.com](https://console.redhat.com)\n\nSyntax of the `filter[name]` query parameters is described in  [Filters documentation](https://github.com/RedHatInsights/patchman-engine/wiki/API-custom-filters)",
        "contact": {},
        "license": {
            "name": "GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "v1.18.52"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/patch/v1/baselines": {
            "get": {
                "summary": "Show me all baselines for all my systems",
                "description": "Show me all baselines for all my systems",
                "operationId": "listBaseline",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit for paging, set -1 to return all",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for paging",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "config"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Find matching text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[id]",
                        "in": "query",
                        "description": "Filter ",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[systems]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tag filter",
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.BaselinesResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            },
            "put": {
                "summary": "Create a baseline for my set of systems",
                "description": "Create a baseline for my set of systems",
                "operationId": "createBaseline",
                "requestBody": {
                    "description": "Request body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/controllers.CreateBaselineRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.CreateBaselineResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ],
                "x-codegen-request-body-name": "body"
            }
        },
...
        "/api/patch/v1/packages/{package_name}": {
            "get": {
                "summary": "Show me metadata of selected package",
                "description": "Show me metadata of selected package",
                "operationId": "LatestPackage",
                "parameters": [
                    {
                        "name": "package_name",
                        "in": "path",
                        "description": "package_name - latest, nevra - exact version",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.PackageDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "controllers.BaselineConfig": {
                "type": "object",
                "properties": {
                    "to_time": {
                        "type": "string",
                        "description": "Filter applicable advisories (updates) by the latest publish time.",
                        "example": "2022-12-31T12:00:00-04:00"
                    }
                }
            },
...
            "controllers.BaselineItem": {
                "type": "object",
                "properties": {
                    "attributes": {
                        "$ref": "#/components/schemas/controllers.BaselineItemAttributes"
                    },
                    "id": {
                        "type": "integer",
                        "description": "Unique baseline id",
                        "example": 10
                    },
                    "type": {
                        "type": "string",
                        "description": "Document type name",
                        "example": "baseline"
                    }
                }
            },
            "controllers.BaselineItemAttributes": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Baseline name",
                        "example": "my-baseline"
                    },
                    "systems": {
                        "type": "integer",
                        "description": "Count of the systems associated with the baseline",
                        "example": 22
                    }
                }
            },
...
            "controllers.BaselinesResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "description": "Baseline items",
                        "items": {
                            "$ref": "#/components/schemas/controllers.BaselineItem"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/controllers.Links"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/controllers.ListMeta"
                    }
                }
            },
            "controllers.CreateBaselineRequest": {
                "type": "object",
                "properties": {
                    "config": {
                        "$ref": "#/components/schemas/controllers.BaselineConfig"
                    },
                    "inventory_ids": {
                        "type": "array",
                        "description": "Inventory IDs list of systems to associate with this baseline (optional).",
                        "items": {
                            "type": "string"
                        }
                    },
                    "name": {
                        "type": "string",
                        "description": "Baseline name (optional)."
                    }
                }
            },
            "controllers.CreateBaselineResponse": {
                "type": "object",
                "properties": {
                    "baseline_id": {
                        "type": "integer",
                        "description": "Updated baseline unique ID, it can not be changed",
                        "example": 1
                    }
                }
            },
...
            "controllers.FilterData": {
                "type": "object",
                "properties": {
                    "op": {
                        "type": "string"
                    },
                    "values": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "controllers.Links": {
                "type": "object",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "/link/to/the/first"
                    },
                    "last": {
                        "type": "string",
                        "example": "/link/to/the/last"
                    },
                    "next": {
                        "type": "string",
                        "example": "/link/to/the/next"
                    },
                    "previous": {
                        "type": "string",
                        "example": "/link/to/the/previous"
                    }
                }
            },
            "controllers.ListMeta": {
                "type": "object",
                "properties": {
                    "filter": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/controllers.FilterData"
                        },
                        "description": "Used filters"
                    },
                    "limit": {
                        "type": "integer",
                        "description": "Used response limit (page size) - pagination",
                        "example": 20
                    },
                    "offset": {
                        "type": "integer",
                        "description": "Used response offset - pagination",
                        "example": 0
                    },
                    "search": {
                        "type": "string",
                        "description": "Used search terms",
                        "example": "kernel"
                    },
                    "sort": {
                        "type": "array",
                        "description": "Used sorting fields",
                        "example": [
                            "name"
                        ],
                        "items": {
                            "type": "string"
                        }
                    },
                    "subtotals": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "description": "Some subtotals used by some endpoints"
                    },
                    "total_items": {
                        "type": "integer",
                        "description": "Total items count to return",
                        "example": 1000
                    }
                }
            },
            "controllers.PackageDetailAttributes": {
                "type": "object",
                "properties": {
                    "advisory_id": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    }
                }
            },
            "controllers.PackageDetailItem": {
                "type": "object",
                "properties": {
                    "attributes": {
                        "$ref": "#/components/schemas/controllers.PackageDetailAttributes"
                    },
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
            "controllers.PackageDetailResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/controllers.PackageDetailItem"
                    }
                }
            },
...
            "utils.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "RhIdentity": {
                "type": "apiKey",
                "name": "x-rh-identity",
                "in": "header"
            }
        }
    }
}